Moved the actual counter example formula construction to the merc_syntax crate, since that is not necessary in merc_refinement.
