+++ crates/refinement/src/refinement.rs
@@ -1,7 +1,6 @@
 use itertools::Itertools;
 use log::debug;
 use log::trace;
-use log::warn;
 use merc_lts::LTS;
 use merc_lts::StateIndex;
 use merc_reduction::Equivalence;
@@ -100,23 +99,23 @@ pub fn refines<L: LTS>(
                 (reduced_lts, StateIndex::new(*spec_block))
             }
             Equivalence::BranchingBisim => {
-                    let (merged_lts, initial_spec) = impl_lts.merge_disjoint(&spec_lts);
-                    let (preprocess_lts, initial_spec, partition) = branching_bisim_sigref(merged_lts, initial_spec, false, timing);
+                let (merged_lts, initial_spec) = impl_lts.merge_disjoint(&spec_lts);
+                let (preprocess_lts, initial_spec, partition) =
+                    branching_bisim_sigref(merged_lts, initial_spec, false, timing);
 
-                    let impl_block = partition.block_number(preprocess_lts.initial_state_index());
-                    let spec_block = partition.block_number(initial_spec);
+                let impl_block = partition.block_number(preprocess_lts.initial_state_index());
+                let spec_block = partition.block_number(initial_spec);
 
-                    if impl_block == spec_block {
-                        // The initial states are already in the same block, so we can skip the refinement check.
-                        debug!(
-                            "Initial states are in the same block after branching bisimulation reduction, skipping refinement check."
-                        );
-                        return (true, None);
-                    }
-
-                    let reduced_lts = quotient_lts_block::<_, true>(&preprocess_lts, &partition);
-                    (reduced_lts, StateIndex::new(*spec_block))
+                if impl_block == spec_block {
+                    // The initial states are already in the same block, so we can skip the refinement check.
+                    debug!(
+                        "Initial states are in the same block after branching bisimulation reduction, skipping refinement check."
+                    );
+                    return (true, None);
                 }
+
+                let reduced_lts = quotient_lts_block::<_, true>(&preprocess_lts, &partition);
+                (reduced_lts, StateIndex::new(*spec_block))
             }
             _ => unimplemented!("Preprocessing for refinement type {refinement:?} has not been implemented yet."),
         }
@@ -220,197 +219,4 @@ pub fn refines<L: LTS>(
             }
         }
     })
-}
-
-#[cfg(test)]
-mod tests {
-    use std::io::Write;
-
-    use merc_vpg::PG;
-    use rand::rngs::StdRng;
-
-    use merc_io::DumpFiles;
-    use merc_lts::mutate_lts;
-    use merc_lts::random_lts_monolithic;
-    use merc_lts::write_aut;
-    use merc_utilities::Timing;
-    use merc_utilities::random_test;
-    use merc_vpg::solve_zielonka;
-    use merc_vpg::translate;
-
-    use crate::ExplorationStrategy;
-    use crate::RefinementType;
-    use crate::generate_formula;
-    use crate::refines;
-
-    #[test]
-    #[cfg_attr(miri, ignore)] // Tests are too slow under miri.
-    fn test_random_trace_refinement() {
-        random_test(100, |rng| {
-            is_refinement_test(
-                "test_random_trace_refinement",
-                rng,
-                RefinementType::Trace,
-                ExplorationStrategy::BFS,
-                false,
-            );
-        });
-    }
-
-    #[test]
-    #[cfg_attr(miri, ignore)] // Tests are too slow under miri.
-    fn test_random_weak_trace_refinement() {
-        random_test(100, |rng| {
-            is_refinement_test(
-                "test_random_weak_trace_refinement",
-                rng,
-                RefinementType::Weaktrace,
-                ExplorationStrategy::BFS,
-                false,
-            );
-        });
-    }
-
-    #[test]
-    #[cfg_attr(miri, ignore)] // Tests are too slow under miri.
-    fn test_random_stable_failures_refinement() {
-        random_test(100, |rng| {
-            is_refinement_test(
-                "test_random_stable_failures_refinement",
-                rng,
-                RefinementType::StableFailures,
-                ExplorationStrategy::BFS,
-                false,
-            );
-        });
-    }
-
-    #[test]
-    #[cfg_attr(miri, ignore)] // Tests are too slow under miri.
-    fn test_random_impossible_futures_refinement() {
-        random_test(100, |rng| {
-            is_refinement_test(
-                "test_random_impossible_futures_refinement",
-                rng,
-                RefinementType::ImpossibleFutures,
-                ExplorationStrategy::BFS,
-                false,
-            );
-        });
-    }
-
-    #[test]
-    #[cfg_attr(miri, ignore)] // Tests are too slow under miri.
-    fn test_random_trace_refinement_preprocessed() {
-        random_test(100, |rng| {
-            is_refinement_test(
-                "test_random_trace_refinement",
-                rng,
-                RefinementType::Trace,
-                ExplorationStrategy::BFS,
-                true,
-            );
-        });
-    }
-
-    #[test]
-    #[cfg_attr(miri, ignore)] // Tests are too slow under miri.
-    fn test_random_weak_trace_refinement_preprocessed() {
-        random_test(100, |rng| {
-            is_refinement_test(
-                "test_random_weak_trace_refinement",
-                rng,
-                RefinementType::Weaktrace,
-                ExplorationStrategy::BFS,
-                true,
-            );
-        });
-    }
-
-    #[test]
-    #[cfg_attr(miri, ignore)] // Tests are too slow under miri.
-    fn test_random_stable_failures_refinement_preprocessed() {
-        random_test(100, |rng| {
-            is_refinement_test(
-                "test_random_stable_failures_refinement",
-                rng,
-                RefinementType::StableFailures,
-                ExplorationStrategy::BFS,
-                true,
-            );
-        });
-    }
-
-    #[test]
-    #[cfg_attr(miri, ignore)] // Tests are too slow under miri.
-    fn test_random_impossible_futures_refinement_preprocessed() {
-        random_test(100, |rng| {
-            is_refinement_test(
-                "test_random_impossible_futures_refinement",
-                rng,
-                RefinementType::ImpossibleFutures,
-                ExplorationStrategy::BFS,
-                true,
-            );
-        });
-    }
-
-    /// Helper function to define a refinement test that can be instantiated for
-    /// the various types.
-    ///
-    /// # Details
-    ///
-    /// Internally requests a counter example to be generated, and checks that
-    /// the counter example is indeed a valid witness for the failure of the
-    /// refinement check.
-    fn is_refinement_test(
-        dump_name: &str,
-        rng: &mut StdRng,
-        refinement: RefinementType,
-        strategy: ExplorationStrategy,
-        preprocess: bool,
-    ) {
-        let mut files = DumpFiles::new(dump_name);
-
-        let spec_lts = random_lts_monolithic(rng, 1000, 5, 3);
-        let impl_lts = mutate_lts(&spec_lts, rng, 100).unwrap();
-
-        files.dump("spec.aut", |w| write_aut(w, &spec_lts)).unwrap();
-        files.dump("impl.aut", |w| write_aut(w, &impl_lts)).unwrap();
-
-        let mut timing = Timing::default();
-        let (result, counter_example) = refines(
-            impl_lts.clone(),
-            spec_lts.clone(),
-            refinement,
-            strategy,
-            preprocess,
-            true,
-            &mut timing,
-        );
-
-        if !result {
-            if let Some(ce) = counter_example {
-                let formula = generate_formula(&ce);
-                println!("Counter example formula: {}", formula);
-
-                files
-                    .dump("counter_example.mcf", |f| Ok(writeln!(f, "{}", formula)?))
-                    .unwrap();
-
-                let impl_pg = translate(&impl_lts, &formula).unwrap();
-                let spec_pg = translate(&spec_lts, &formula).unwrap();
-
-                let (impl_solution, _) = solve_zielonka(&impl_pg, false);
-                let (spec_solution, _) = solve_zielonka(&spec_pg, false);
-
-                assert!(
-                    impl_solution[impl_pg.initial_vertex()] != spec_solution[spec_pg.initial_vertex()],
-                    "Refinement returned false, but the counter example is not distinguishing."
-                );
-            } else {
-                panic!("Expected a counter example.");
-            }
-        }
-    }
 }
