+++ crates/syntax/Cargo.toml
@@ -14,6 +14,7 @@ version.workspace = true
 
 [dependencies]
 merc_lts.workspace = true
+merc_refinement.workspace = true
 merc_pest_consume.workspace = true
 merc_utilities.workspace = true
