+++ tools/lts/Cargo.toml
@@ -9,8 +9,9 @@ rust-version.workspace = true
 merc_io.workspace = true
 merc_ldd.workspace = true
 merc_lts = { workspace = true, features = ["clap"] }
-merc_refinement = { workspace = true, features = ["clap"] }
 merc_reduction = { workspace = true, features = ["clap"] }
+merc_refinement = { workspace = true, features = ["clap"] }
+merc_syntax.workspace = true
 merc_tools.workspace = true
 merc_unsafety.workspace = true
 merc_utilities.workspace = true
